Move app components in sub-folder
