+++ src/components/App/app-wrapper.tsx
@@ -28,15 +28,15 @@ import {
 import { IntlProvider } from 'react-intl';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider, useSelector } from 'react-redux';
-import { SupportedLanguages } from '../utils/language';
-import messages_en from '../translations/en.json';
-import messages_fr from '../translations/fr.json';
-import messages_plugins_en from '../plugins/translations/en.json';
-import messages_plugins_fr from '../plugins/translations/fr.json';
-import { store } from '../redux/store';
-import { PARAM_THEME } from '../utils/config-params';
+import { SupportedLanguages } from '../../utils/language';
+import messages_en from '../../translations/en.json';
+import messages_fr from '../../translations/fr.json';
+import messages_plugins_en from '../../plugins/translations/en.json';
+import messages_plugins_fr from '../../plugins/translations/fr.json';
+import { store } from '../../redux/store';
+import { PARAM_THEME } from '../../utils/config-params';
 import { IntlConfig } from 'react-intl/src/types';
-import { AppState } from '../redux/reducer';
+import { AppState } from '../../redux/reducer';
 
 const lightTheme: Theme = createTheme({
     palette: {
