+++ src/components/App/app.test.tsx
@@ -7,7 +7,7 @@ import { IntlProvider } from 'react-intl';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import App from './app';
-import { store } from '../redux/store';
+import { store } from '../../redux/store';
 import {
     createTheme,
     StyledEngineProvider,
