+++ src/components/App/index.ts
@@ -1,0 +1,1 @@
+export { default as AppWrapper } from './app-wrapper';
